Added support for doubled-lines drawing.
Classical bits (and more generally classical operations) are depicted using
a doubled-line, so I added some support for this. Some changes still need to
be done, for example for the measurement (the line between the gate and the
classical register is still a simple line), but this will be done later.