Font-size auto-resize and register names
The font size is now automatically resized to fill as much as possible
the gates.
Moreover, the registers' names are drawn before each register line.
A possible improvement would be to provide to the user a way to choose
where he/she wants to draw the names: left, right, both?