First try for the bit ordering.
The bit ordering should work in _draw_json_circuit. I did not test
it for the moment, but I will.
The main problem now is to determine how to take the bit ordering
as input to qasm2svg or even qasm2image. The problem comes from the
fact that the bits of the given QASM are only known during qasm2svg,
and so before qasm2svg we don't have any clue on which bit is present
or not.
One way to solve this would be to define some textual descriptors
like "reverse", "default", "interleaved", "bigendian", "littleendian",
etc... but I find this method a little bit too rigid: what if the
user want 1 qubit in little endian and all the others in interleaved
big endian?