qasm2svg outputs the dimensions of the returned SVG.
The qasm2svg function was only returning a string, i.e. the circuit
in SVG format. Now, the function behaviour does not change by default
but an additional flag can be given as parameter. If the flag is set
to True, the function will return the SVG representation and a tuple
containing the width and the height of the SVG.