Fixed a bug with barriers.
Barrier were correctly drawn on the SVG representation but the
dimensions were invalid. This was due to the fact that the
computation of the width did not take into account the specific
case of the barrier gate (which is not drawn in the SVG, so we don't
have to reserve a place for its representation).
Now the _update_data_structure procedure is used in both the computation
of SVG dimensions and the drawing process, so the computed dimensions
should always agree.